Struct bounded_int::BoundedInt

pub struct BoundedInt<const LOW: i32, const HIGH: i32>(pub(crate) i32);

An integer bound between two points, inclusive on both ends.

Examples

use std::ops::{Add, Sub};
use bounded_int::{BoundedInt, BoundedIntError};

let b = BoundedInt::<-10, 10>::new(0).unwrap();

assert_eq!(b.add(5).unwrap(), BoundedInt::<-10, 10>::new(5).unwrap());

assert_eq!(b.sub(5).unwrap(), BoundedInt::<-10, 10>::new(-5).unwrap());

assert_eq!(
    b.saturating_add(100),
    BoundedInt::<-10, 10>::new(10).unwrap()
);

assert_eq!(
    b.saturating_sub(100),
    BoundedInt::<-10, 10>::new(-10).unwrap()
);

let invalid = BoundedInt::<-10, 10>::new(100);

assert_eq!(invalid, Err(BoundedIntError::TooHigh(100)));

Implementations

impl<const LOW: i32, const HIGH: i32> BoundedInt<{ LOW }, { HIGH }>

pub const LOW: i32

The lower bounds of the BoundedInt, inclusive.

pub const HIGH: i32

The upper bounds of the BoundedInt, inclusive.

pub fn new_clamped(n: i32) -> Self

Returns the input as a BoundedInt, clamped at the bounds.

Examples

use bounded_int::{BoundedInt, BoundedIntError};

let clamped = BoundedInt::<20, 30>::new_clamped(15);

assert_eq!(clamped, BoundedInt::<20, 30>::new(20).unwrap());

pub fn new(n: i32) -> Result<Self, BoundedIntError>

Attempts to create a new BoundedInt with the provided value.

Errors

This function will error if the supplied value doesn’t fit within the bounds required.

pub const fn as_unbounded(self) -> i32

Returns the inner value of the BoundedInt.

pub const fn range_to(self, to: Self) -> BoundedIntRange<{ LOW }, { HIGH }>

Returns a BoundedIntRange from self to to.

If to <= self, the iterator will be empty.

Examples

use bounded_int::BoundedInt;

let start = BoundedInt::<10, 15>::new(10).unwrap();
let end = BoundedInt::<10, 15>::new(15).unwrap();

assert_eq!(
    start.range_to(end).collect::<Vec<_>>(),
    &[
        BoundedInt::<10, 15>::new(10).unwrap(),
        BoundedInt::<10, 15>::new(11).unwrap(),
        BoundedInt::<10, 15>::new(12).unwrap(),
        BoundedInt::<10, 15>::new(13).unwrap(),
        BoundedInt::<10, 15>::new(14).unwrap(),
    ]
);

pub const fn range_from(self, from: Self) -> BoundedIntRange<{ LOW }, { HIGH }>

Returns a BoundedIntRange from from to self.

If self <= from, the iterator will be empty.

Examples

use bounded_int::BoundedInt;

let start = BoundedInt::<10, 15>::new(10).unwrap();
let end = BoundedInt::<10, 15>::new(15).unwrap();

assert_eq!(
    end.range_from(start).collect::<Vec<_>>(),
    &[
        BoundedInt::<10, 15>::new(10).unwrap(),
        BoundedInt::<10, 15>::new(11).unwrap(),
        BoundedInt::<10, 15>::new(12).unwrap(),
        BoundedInt::<10, 15>::new(13).unwrap(),
        BoundedInt::<10, 15>::new(14).unwrap(),
    ]
);

pub const fn range_to_inclusive(
self,
to: Self
) -> BoundedIntRangeInclusive<{ LOW }, { HIGH }>

Returns a BoundedIntRangeInclusive from self to to.

If to <= self, the iterator will be empty. If to == self, the iterator will produce one item equal to self.

Examples

use bounded_int::BoundedInt;

let start = BoundedInt::<10, 15>::new(10).unwrap();
let end = BoundedInt::<10, 15>::new(15).unwrap();

assert_eq!(
    start.range_to_inclusive(end).collect::<Vec<_>>(),
    &[
        BoundedInt::<10, 15>::new(10).unwrap(),
        BoundedInt::<10, 15>::new(11).unwrap(),
        BoundedInt::<10, 15>::new(12).unwrap(),
        BoundedInt::<10, 15>::new(13).unwrap(),
        BoundedInt::<10, 15>::new(14).unwrap(),
        BoundedInt::<10, 15>::new(15).unwrap(),
    ]
);

pub const fn range_from_inclusive(
self,
from: Self
) -> BoundedIntRangeInclusive<{ LOW }, { HIGH }>

Returns a BoundedIntRangeInclusive from from to self.

If self < from, the iterator will be empty. If self == from, the iterator will produce one item equal to self.

Examples

use bounded_int::BoundedInt;

let start = BoundedInt::<10, 15>::new(10).unwrap();
let end = BoundedInt::<10, 15>::new(15).unwrap();

assert_eq!(
    end.range_from_inclusive(start).collect::<Vec<_>>(),
    &[
        BoundedInt::<10, 15>::new(10).unwrap(),
        BoundedInt::<10, 15>::new(11).unwrap(),
        BoundedInt::<10, 15>::new(12).unwrap(),
        BoundedInt::<10, 15>::new(13).unwrap(),
        BoundedInt::<10, 15>::new(14).unwrap(),
        BoundedInt::<10, 15>::new(15).unwrap(),
    ]
);

pub fn expand_upper<const HIGHER: i32>(self) -> BoundedInt<{ LOW }, { HIGHER }>

Raises the upper bounds of the BoundedInt to HIGHER.

Examples

use bounded_int::BoundedInt;

let small_bounds = BoundedInt::<10, 15>::new(10).unwrap();

fn requires_larger_bounds(b: BoundedInt::<10, 20>) { }

// type inference makes this very simple to call
requires_larger_bounds(small_bounds.expand_upper());

pub fn expand_lower<const LOWER: i32>(self) -> BoundedInt<{ LOWER }, { HIGH }>

Increases the lower bounds of the BoundedInt to LOWER.

Examples

use bounded_int::BoundedInt;

let small_bounds = BoundedInt::<10, 15>::new(10).unwrap();

fn requires_larger_bounds(b: BoundedInt::<5, 15>) { }

// type inference makes this very simple to call
requires_larger_bounds(small_bounds.expand_lower());

pub fn expand_bounds<const LOWER: i32, const HIGHER: i32>(
self
) -> BoundedInt<{ LOWER }, { HIGHER }>

Expands the bounds of the BoundedInt to LOWER and HIGHER.

Examples

use bounded_int::BoundedInt;

let small_bounds = BoundedInt::<10, 15>::new(10).unwrap();

fn requires_larger_bounds(b: BoundedInt::<5, 20>) { }

// type inference makes this very simple to call
requires_larger_bounds(small_bounds.expand_bounds());

pub fn saturating_sub(self, rhs: i32) -> Self

Performs subtraction that saturates at the numeric bounds instead of overflowing.

pub fn saturating_add(self, rhs: i32) -> Self

Performs addition that saturates at the numeric bounds instead of overflowing.