Crate bounded_int

Leverages const generics to enforce compile time bounds on integers.

Note: due to current limitations in Rust, the BoundedInt type uses an i32 behind the scenes. This could be wasteful if you limit the bounds to smaller than, say, an i16, or may not be enough if you need a number larger than an i32 can hold.

Re-exports

pub use crate::iter::BoundedIntRange;

pub use crate::iter::BoundedIntRangeInclusive;

Modules

iter
Contains the Iterator implementations for the range types.

ops
Contains the various std::ops trait implementations for BoundedInt.

Structs

BoundedInt
An integer bound between two points, inclusive on both ends.

Enums

BoundedIntError
Error returned when trying to convert an i32 to a BoundedInt.